requirements.txt
Flask==3.0.3
Flask-Compress==1.17
SQLAlchemy==2.0.36
python-dotenv==1.0.1
reportlab==4.2.5
qrcode==8.0
Pillow==11.0.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
stripe==11.2.0
APScheduler==3.10.4
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
joblib==1.4.2
openpyxl==3.1.5
requests==2.32.3
psycopg2-binary==2.9.10